Added JUnit method testInverse to test method Inverse in class Matrix

+++ JavaVersion/src/classFiles/Matrix.java
@@ -73,7 +73,7 @@ public class Matrix {
 	}
 	
 	// adds an element to the matrix
-	public void addElement(int element, int r, int c) {
+	public void addElement(double element, int r, int c) {
 		if( (r >= 0) && (c >= 0) && (r < n) && (c < n) ) {
 			mdata[r][c] = element;
 		}
